Refactor wrappers to single file to allow external use
Other repositories can sync this file as-is (i.e. via dotnet-file sync) and reuse the wrappers to also get git-less usage of GCM in their own scenarios.
